# src/backend/api/swagger.json
{
	"openapi": "3.0.0",
	"components": {
		"examples": {},
		"headers": {},
		"parameters": {},
		"requestBodies": {},
		"responses": {},
		"schemas": {
			"TestCase": {
				"properties": {
					"testcaseId": {
						"type": "number",
						"format": "double"
					},
					"input": {
						"type": "string"
					},
					"expectedOutput": {
						"type": "string"
					},
					"timeLimitMs": {
						"type": "number",
						"format": "double"
					},
					"memoryLimitMb": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"testcaseId",
					"input",
					"expectedOutput",
					"timeLimitMs",
					"memoryLimitMb"
				],
				"type": "object",
				"additionalProperties": false
			},
			"NotFoundError": {
				"properties": {
					"name": {
						"type": "string"
					},
					"message": {
						"type": "string"
					},
					"stack": {
						"type": "string"
					},
					"statusCode": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"name",
					"message",
					"statusCode"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ValidateError": {
				"properties": {
					"message": {
						"type": "string",
						"enum": [
							"Validation failed"
						],
						"nullable": false
					},
					"details": {
						"properties": {},
						"additionalProperties": {},
						"type": "object"
					}
				},
				"required": [
					"message",
					"details"
				],
				"type": "object",
				"additionalProperties": false
			},
			"CreateTestCaseParams": {
				"properties": {
					"input": {
						"type": "string",
						"minLength": 1
					},
					"expectedOutput": {
						"type": "string",
						"minLength": 1
					},
					"timeLimitMs": {
						"type": "number",
						"format": "double",
						"minimum": 100,
						"maximum": 10000
					},
					"memoryLimitMb": {
						"type": "number",
						"format": "double",
						"minimum": 16,
						"maximum": 1024
					}
				},
				"required": [
					"input",
					"expectedOutput",
					"timeLimitMs",
					"memoryLimitMb"
				],
				"type": "object",
				"additionalProperties": false
			},
			"SubmissionResultDto": {
				"properties": {
					"status": {
						"type": "string"
					},
					"output": {
						"type": "string"
					},
					"expectedOutput": {
						"type": "string"
					},
					"runtimeMs": {
						"type": "number",
						"format": "double"
					},
					"memoryKb": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"status",
					"runtimeMs",
					"memoryKb"
				],
				"type": "object",
				"additionalProperties": false
			},
			"RunCodeResponseDto": {
				"properties": {
					"status": {
						"type": "string"
					},
					"results": {
						"items": {
							"$ref": "#/components/schemas/SubmissionResultDto"
						},
						"type": "array"
					}
				},
				"required": [
					"status",
					"results"
				],
				"type": "object",
				"additionalProperties": false
			},
			"RunCodeDto": {
				"properties": {
					"code": {
						"type": "string",
						"minLength": 1
					},
					"languageId": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"code",
					"languageId"
				],
				"type": "object",
				"additionalProperties": false
			},
			"SubmitCodeResponseDto": {
				"properties": {
					"submissionId": {
						"type": "number",
						"format": "double"
					},
					"overallStatus": {
						"type": "string"
					},
					"results": {
						"items": {
							"$ref": "#/components/schemas/SubmissionResultDto"
						},
						"type": "array"
					}
				},
				"required": [
					"submissionId",
					"overallStatus",
					"results"
				],
				"type": "object",
				"additionalProperties": false
			},
			"SubmitCodeDto": {
				"properties": {
					"code": {
						"type": "string",
						"minLength": 1
					},
					"languageId": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"code",
					"languageId"
				],
				"type": "object",
				"additionalProperties": false
			},
			"SubmissionListItemDto": {
				"properties": {
					"submissionId": {
						"type": "number",
						"format": "double"
					},
					"code": {
						"type": "string"
					},
					"languageId": {
						"type": "number",
						"format": "double"
					},
					"status": {
						"type": "string"
					},
					"submittedAt": {
						"type": "string"
					}
				},
				"required": [
					"submissionId",
					"code",
					"languageId",
					"status",
					"submittedAt"
				],
				"type": "object",
				"additionalProperties": false
			},
			"SubmissionDetailDto": {
				"properties": {
					"submissionId": {
						"type": "number",
						"format": "double"
					},
					"code": {
						"type": "string"
					},
					"languageId": {
						"type": "number",
						"format": "double"
					},
					"status": {
						"type": "string"
					},
					"submittedAt": {
						"type": "string"
					},
					"results": {
						"items": {
							"$ref": "#/components/schemas/SubmissionResultDto"
						},
						"type": "array"
					}
				},
				"required": [
					"submissionId",
					"code",
					"languageId",
					"status",
					"submittedAt",
					"results"
				],
				"type": "object",
				"additionalProperties": false
			},
			"_36_Enums.Difficulty": {
				"type": "string",
				"enum": [
					"EASY",
					"MEDIUM",
					"HARD"
				]
			},
			"Difficulty": {
				"$ref": "#/components/schemas/_36_Enums.Difficulty"
			},
			"CompletionState": {
				"type": "string",
				"enum": [
					"Completed",
					"Attempted",
					"Unattempted"
				]
			},
			"ProblemSummary": {
				"properties": {
					"problemId": {
						"type": "number",
						"format": "double"
					},
					"title": {
						"type": "string"
					},
					"difficulty": {
						"$ref": "#/components/schemas/Difficulty"
					},
					"completionState": {
						"$ref": "#/components/schemas/CompletionState"
					}
				},
				"required": [
					"problemId",
					"title",
					"difficulty",
					"completionState"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ProblemDetails": {
				"properties": {
					"problemId": {
						"type": "number",
						"format": "double"
					},
					"title": {
						"type": "string"
					},
					"difficulty": {
						"$ref": "#/components/schemas/Difficulty"
					},
					"description": {
						"type": "string"
					},
					"createdAt": {
						"type": "string"
					}
				},
				"required": [
					"problemId",
					"title",
					"difficulty",
					"description",
					"createdAt"
				],
				"type": "object",
				"additionalProperties": false
			},
			"CreateProblemParams": {
				"properties": {
					"title": {
						"type": "string",
						"minLength": 5,
						"maxLength": 100
					},
					"description": {
						"type": "string",
						"minLength": 10,
						"maxLength": 2000
					},
					"difficulty": {
						"$ref": "#/components/schemas/Difficulty"
					}
				},
				"required": [
					"title",
					"description",
					"difficulty"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UpdateProblemParams": {
				"properties": {
					"title": {
						"type": "string",
						"minLength": 5,
						"maxLength": 100
					},
					"description": {
						"type": "string",
						"minLength": 10,
						"maxLength": 2000
					},
					"difficulty": {
						"$ref": "#/components/schemas/Difficulty"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
			"LanguageDto": {
				"properties": {
					"languageId": {
						"type": "number",
						"format": "double"
					},
					"name": {
						"type": "string"
					},
					"suffix": {
						"type": "string"
					},
					"version": {
						"type": "string",
						"nullable": true
					},
					"compilerCmd": {
						"type": "string",
						"nullable": true
					},
					"runtimeCmd": {
						"type": "string"
					},
					"compile_command": {
						"type": "string",
						"nullable": true,
						"deprecated": true
					},
					"run_command": {
						"type": "string",
						"deprecated": true
					},
					"isDefault": {
						"type": "boolean"
					}
				},
				"required": [
					"languageId",
					"name",
					"suffix",
					"version",
					"compilerCmd",
					"runtimeCmd",
					"isDefault"
				],
				"type": "object",
				"additionalProperties": false
			},
			"CreateLanguageRequestDto": {
				"properties": {
					"name": {
						"type": "string",
						"description": "Language name (e.g., \"Python\")",
						"minLength": 1,
						"maxLength": 50
					},
					"runtimeCmd": {
						"type": "string",
						"description": "Command used to run the compiled/interpreted program",
						"minLength": 1
					},
					"run_command": {
						"type": "string",
						"deprecated": true
					},
					"compile_command": {
						"type": "string",
						"nullable": true,
						"deprecated": true
					},
					"compilerCmd": {
						"type": "string",
						"nullable": true,
						"description": "Optional compile command (null for interpreted languages)"
					},
					"version": {
						"type": "string",
						"nullable": true
					},
					"suffix": {
						"type": "string",
						"minLength": 1
					}
				},
				"required": [
					"name",
					"runtimeCmd",
					"suffix"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UpdateLanguageRequestDto": {
				"properties": {
					"name": {
						"type": "string"
					},
					"runtimeCmd": {
						"type": "string"
					},
					"run_command": {
						"type": "string",
						"deprecated": true
					},
					"compile_command": {
						"type": "string",
						"nullable": true,
						"deprecated": true
					},
					"compilerCmd": {
						"type": "string",
						"nullable": true
					},
					"version": {
						"type": "string",
						"nullable": true
					},
					"suffix": {
						"type": "string",
						"minLength": 1
					}
				},
				"type": "object",
				"additionalProperties": false
			},
			"ForbiddenError": {
				"description": "ForbiddenError — represents an HTTP 403 error.\n\nThrow this when the client tries to perform an action that is not allowed,\ne.g., deleting the default programming language.",
				"properties": {
					"name": {
						"type": "string"
					},
					"message": {
						"type": "string"
					},
					"stack": {
						"type": "string"
					},
					"statusCode": {
						"type": "number",
						"format": "double",
						"description": "HTTP status code to be picked up by the global error handler",
						"default": 403
					}
				},
				"required": [
					"name",
					"message"
				],
				"type": "object",
				"additionalProperties": false
			}
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "algo-platform",
		"version": "1.0.0",
		"description": "My Electron application description",
		"license": {
			"name": "MIT"
		},
		"contact": {
			"name": "Chen Lu (lupanpan)",
			"email": "clu396@aucklanduni.ac.nz"
		}
	},
	"paths": {
		"/problems/{problemId}/testcases": {
			"get": {
				"operationId": "GetTestCases",
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/TestCase"
									},
									"type": "array"
								}
							}
						}
					},
					"404": {
						"description": "Problem not found",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/NotFoundError"
								}
							}
						}
					}
				},
				"description": "Retrieves all test cases for a given problem.",
				"tags": [
					"TestCases"
				],
				"security": [],
				"parameters": [
					{
						"description": "The ID of the problem.",
						"in": "path",
						"name": "problemId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"post": {
				"operationId": "CreateTestCase",
				"responses": {
					"201": {
						"description": "Created",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/TestCase"
								}
							}
						}
					},
					"404": {
						"description": "Problem not found",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/NotFoundError"
								}
							}
						}
					},
					"422": {
						"description": "Validation Failed",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ValidateError"
								}
							}
						}
					}
				},
				"description": "Creates a new test case for a given problem.",
				"tags": [
					"TestCases"
				],
				"security": [],
				"parameters": [
					{
						"description": "The ID of the problem.",
						"in": "path",
						"name": "problemId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"description": "The test case data.",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateTestCaseParams",
								"description": "The test case data."
							}
						}
					}
				}
			}
		},
		"/problems/{problemId}/testcases/{testcaseId}": {
			"delete": {
				"operationId": "DeleteTestCase",
				"responses": {
					"204": {
						"description": "No Content"
					},
					"404": {
						"description": "Problem not found",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/NotFoundError"
								}
							}
						}
					}
				},
				"description": "Deletes a test case for a given problem.",
				"tags": [
					"TestCases"
				],
				"security": [],
				"parameters": [
					{
						"description": "The ID of the problem.",
						"in": "path",
						"name": "problemId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"description": "The ID of the test case to delete.",
						"in": "path",
						"name": "testcaseId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/problems/{problemId}/run": {
			"post": {
				"operationId": "RunCode",
				"responses": {
					"200": {
						"description": "Code ran successfully",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/RunCodeResponseDto"
								}
							}
						}
					},
					"404": {
						"description": "Problem, language or test case not found",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/NotFoundError"
								}
							}
						}
					},
					"422": {
						"description": "Validation Failed",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ValidateError"
								}
							}
						}
					}
				},
				"description": "Run code against a subset of test cases without saving the submission",
				"tags": [
					"Submissions"
				],
				"security": [],
				"parameters": [
					{
						"description": "The problem ID",
						"in": "path",
						"name": "problemId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"description": "The code and language ID",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/RunCodeDto",
								"description": "The code and language ID"
							}
						}
					}
				}
			}
		},
		"/problems/{problemId}/submit": {
			"post": {
				"operationId": "SubmitCode",
				"responses": {
					"201": {
						"description": "Code submitted successfully",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/SubmitCodeResponseDto"
								}
							}
						}
					},
					"404": {
						"description": "Problem, language or test case not found",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/NotFoundError"
								}
							}
						}
					},
					"422": {
						"description": "Validation Failed",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ValidateError"
								}
							}
						}
					}
				},
				"description": "Submit code for full evaluation and save the submission",
				"tags": [
					"Submissions"
				],
				"security": [],
				"parameters": [
					{
						"description": "The problem ID",
						"in": "path",
						"name": "problemId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"description": "The code and language ID",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/SubmitCodeDto",
								"description": "The code and language ID"
							}
						}
					}
				}
			}
		},
		"/problems/{problemId}/submissions": {
			"get": {
				"operationId": "GetSubmissionsByProblem",
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/SubmissionListItemDto"
									},
									"type": "array"
								}
							}
						}
					},
					"404": {
						"description": "Problem not found",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/NotFoundError"
								}
							}
						}
					}
				},
				"description": "List submissions for a problem",
				"tags": [
					"Submissions"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "problemId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/problems/{problemId}/submissions/{submissionId}": {
			"get": {
				"operationId": "GetSubmissionByProblem",
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/SubmissionDetailDto"
								}
							}
						}
					},
					"404": {
						"description": "Submission not found",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/NotFoundError"
								}
							}
						}
					}
				},
				"description": "Get a specific submission for a problem",
				"tags": [
					"Submissions"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "problemId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "path",
						"name": "submissionId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/problems": {
			"get": {
				"operationId": "GetAllProblems",
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/ProblemSummary"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"description": "Retrieves a list of all problems with summary information.",
				"tags": [
					"Problems"
				],
				"security": [],
				"parameters": []
			},
			"post": {
				"operationId": "CreateProblem",
				"responses": {
					"201": {
						"description": "Created",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ProblemDetails"
								}
							}
						}
					},
					"422": {
						"description": "Validation Failed",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ValidateError"
								}
							}
						}
					}
				},
				"description": "Creates a new problem.",
				"tags": [
					"Problems"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"description": "The data required to create the problem.",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateProblemParams",
								"description": "The data required to create the problem."
							}
						}
					}
				}
			}
		},
		"/problems/{problemId}": {
			"get": {
				"operationId": "GetProblem",
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ProblemDetails"
								}
							}
						}
					},
					"404": {
						"description": "Problem not found",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/NotFoundError"
								}
							}
						}
					}
				},
				"description": "Retrieves the detailed information of a specific problem by its ID.",
				"tags": [
					"Problems"
				],
				"security": [],
				"parameters": [
					{
						"description": "The ID of the problem.",
						"in": "path",
						"name": "problemId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"put": {
				"operationId": "UpdateProblem",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ProblemDetails"
								}
							}
						}
					},
					"404": {
						"description": "Problem not found",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/NotFoundError"
								}
							}
						}
					},
					"422": {
						"description": "Validation Failed",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ValidateError"
								}
							}
						}
					}
				},
				"description": "Updates an existing problem.",
				"tags": [
					"Problems"
				],
				"security": [],
				"parameters": [
					{
						"description": "The ID of the problem to update.",
						"in": "path",
						"name": "problemId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"description": "The updated problem data.",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdateProblemParams",
								"description": "The updated problem data."
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteProblem",
				"responses": {
					"204": {
						"description": "No Content"
					}
				},
				"description": "Deletes a problem by its ID.",
				"tags": [
					"Problems"
				],
				"security": [],
				"parameters": [
					{
						"description": "The ID of the problem to delete.",
						"in": "path",
						"name": "problemId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/languages": {
			"get": {
				"operationId": "GetLanguages",
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/LanguageDto"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"description": "List all programming languages.",
				"tags": [
					"ProgrammingLanguage"
				],
				"security": [],
				"parameters": []
			},
			"post": {
				"operationId": "Create",
				"responses": {
					"201": {
						"description": "Created",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/LanguageDto"
								}
							}
						}
					},
					"422": {
						"description": "Validation Failed",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ValidateError"
								}
							}
						}
					}
				},
				"description": "Create a new programming language.",
				"tags": [
					"ProgrammingLanguage"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"description": "Language information",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateLanguageRequestDto",
								"description": "Language information"
							}
						}
					}
				}
			}
		},
		"/languages/{id}": {
			"get": {
				"operationId": "Get",
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/LanguageDto"
								}
							}
						}
					},
					"404": {
						"description": "Language not found",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/NotFoundError"
								}
							}
						}
					}
				},
				"description": "Get a programming language by its ID.",
				"tags": [
					"ProgrammingLanguage"
				],
				"security": [],
				"parameters": [
					{
						"description": "Language ID",
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"put": {
				"operationId": "Update",
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/LanguageDto"
								}
							}
						}
					},
					"404": {
						"description": "Language not found",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/NotFoundError"
								}
							}
						}
					},
					"422": {
						"description": "Validation Failed",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ValidateError"
								}
							}
						}
					}
				},
				"description": "Update an existing programming language.",
				"tags": [
					"ProgrammingLanguage"
				],
				"security": [],
				"parameters": [
					{
						"description": "Language ID",
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"description": "Updated language information",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdateLanguageRequestDto",
								"description": "Updated language information"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "Remove",
				"responses": {
					"204": {
						"description": "No Content"
					},
					"403": {
						"description": "Default language cannot be deleted",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ForbiddenError"
								}
							}
						}
					},
					"404": {
						"description": "Language not found",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/NotFoundError"
								}
							}
						}
					}
				},
				"description": "Delete a programming language.",
				"tags": [
					"ProgrammingLanguage"
				],
				"security": [],
				"parameters": [
					{
						"description": "Language ID",
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		}
	},
	"servers": [
		{
			"url": "/"
		}
	]
}